Refactor Quilkin's data pipeline into distinct types.
