+++ Cargo.toml
@@ -63,6 +63,9 @@ tokio-stream = "0.1.2"
 tonic = "0.5.0"
 uuid = {version = "0.8.1", default-features = false, features = ["v4"]}
 thiserror = "1.0.25"
+once_cell = "1.8.0"
+futures = "0.3.15"
+arc-swap = "1.3.0"
 
 [dev-dependencies]
 reqwest = "0.11.0"
