+++ src/proxy/resource_manager.rs
